+++ Demo/VComponentsDemo/Models/HomeRow.swift
@@ -10,7 +10,7 @@ import SwiftUI
 // MARK: - Home Row
 enum HomeRow: Int, DemoableRow {
     // MARK: Cases
-    case primaryButton, secondaryButton, squareButton, plainButton, navigationLink, link
+    case primaryButton, secondaryButton, roundedButton, plainButton, navigationLink, link
     case toggle, checkBox, radioButton
     case segmentedPicker, wheelPicker
     case stepper, slider,  rangeSlider
@@ -26,7 +26,7 @@ enum HomeRow: Int, DemoableRow {
     static var sections: [DemoSection<Self>] {
         [
             .init(id: 0, title: "Buttons", rows: [
-                .primaryButton, .secondaryButton, .squareButton, .plainButton,
+                .primaryButton, .secondaryButton, .roundedButton, .plainButton,
                 .navigationLink, .link
             ]),
             .init(id: 1, title: "State Pickers", rows: [.toggle, .checkBox, .radioButton]),
@@ -46,7 +46,7 @@ enum HomeRow: Int, DemoableRow {
         switch self {
         case .primaryButton: return VPrimaryButtonDemoView.navBarTitle
         case .secondaryButton: return VSecondaryButtonDemoView.navBarTitle
-        case .squareButton: return VSquareButtonDemoView.navBarTitle
+        case .roundedButton: return VRoundedButtonDemoView.navBarTitle
         case .plainButton: return VPlainButtonDemoView.navBarTitle
         case .navigationLink: return VNavigationLinkDemoView.navBarTitle
         case .link: return VLinkDemoView.navBarTitle
@@ -91,7 +91,7 @@ enum HomeRow: Int, DemoableRow {
         switch self {
         case .primaryButton: VPrimaryButtonDemoView()
         case .secondaryButton: VSecondaryButtonDemoView()
-        case .squareButton: VSquareButtonDemoView()
+        case .roundedButton: VRoundedButtonDemoView()
         case .plainButton: VPlainButtonDemoView()
         case .navigationLink: VNavigationLinkDemoView()
         case .link: VLinkDemoView()