+++ Sources/VComponents/Components/Buttons/Square Button/VRoundedButtonUIModel.swift
@@ -1,5 +1,5 @@
 //
-//  VSquareButtonUIModel.swift
+//  VRoundedButtonUIModel.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 19.12.20.
@@ -8,9 +8,9 @@
 import SwiftUI
 import VCore
 
-// MARK: - V Square Button UI Model
+// MARK: - V Rounded Button UI Model
 /// Model that describes UI.
-public struct VSquareButtonUIModel {
+public struct VRoundedButtonUIModel {
     // MARK: Properties
     fileprivate static let primaryButtonReference: VPrimaryButtonUIModel = .init()