+++ Sources/VComponents/Components/Buttons/Square Button/VSquareButtonState.swift
@@ -1,5 +1,5 @@
 //
-//  VSquareButtonState.swift
+//  VRoundedButtonState.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 19.12.20.
@@ -8,5 +8,5 @@
 import SwiftUI
 import VCore
 
-// MARK: - V Square Button Internal State
-typealias VSquareButtonInternalState = GenericState_EnabledPressedDisabled
+// MARK: - V Rounded Button Internal State
+typealias VRoundedButtonInternalState = GenericState_EnabledPressedDisabled