+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheet.swift
@@ -217,7 +217,7 @@ struct VBottomSheet<HeaderLabel, Content>: View
     }
 
     private var closeButton: some View {
-        VSquareButton.close(
+        VRoundedButton.close(
             uiModel: uiModel.closeButtonSubUIModel,
             action: animateOut
         )